+++ weed/s3api/s3api_bucket_handlers.go
@@ -667,41 +667,3 @@ func (s3a *S3ApiServer) GetBucketVersioningHandler(w http.ResponseWriter, r *htt
 		},
 	})
 }
-
-// GetBucketTaggingHandler Returns the tag set associated with the bucket
-// https://docs.aws.amazon.com/AmazonS3/latest/API/API_GetBucketTagging.html
-func (s3a *S3ApiServer) GetBucketTaggingHandler(w http.ResponseWriter, r *http.Request) {
-	bucket, _ := s3_constants.GetBucketAndObject(r)
-	glog.V(3).Infof("GetBucketTagging %s", bucket)
-
-	if err := s3a.checkBucket(r, bucket); err != s3err.ErrNone {
-		s3err.WriteErrorResponse(w, r, err)
-		return
-	}
-
-	s3err.WriteErrorResponse(w, r, s3err.ErrNoSuchTagSet)
-}
-
-func (s3a *S3ApiServer) PutBucketTaggingHandler(w http.ResponseWriter, r *http.Request) {
-	bucket, _ := s3_constants.GetBucketAndObject(r)
-	glog.V(3).Infof("PutBucketTagging %s", bucket)
-
-	if err := s3a.checkBucket(r, bucket); err != s3err.ErrNone {
-		s3err.WriteErrorResponse(w, r, err)
-		return
-	}
-
-	s3err.WriteErrorResponse(w, r, s3err.ErrNotImplemented)
-}
-
-func (s3a *S3ApiServer) DeleteBucketTaggingHandler(w http.ResponseWriter, r *http.Request) {
-	bucket, _ := s3_constants.GetBucketAndObject(r)
-	glog.V(3).Infof("DeleteBucketTagging %s", bucket)
-
-	if err := s3a.checkBucket(r, bucket); err != s3err.ErrNone {
-		s3err.WriteErrorResponse(w, r, err)
-		return
-	}
-
-	s3err.WriteErrorResponse(w, r, s3err.ErrNotImplemented)
-}
